// src/components/UserContibution/types.ts
export interface DisplayContributionDTO  {
  assignee?: string;
  assigneeId?: string;
  createdAt?: number;
  createdBy?: string;
  endsOn?: number;
  id?: string;
  isNoteworthy?: boolean;
  percentCompleted?: number;
  priority?: string;
  startedOn?: number;
  status?: string;
  title?: string;
  type?: string;
  updatedAt?: number;
};